+++ src/dftracer/utils/indexer/indexer_c.cpp
@@ -17,7 +17,7 @@ static Indexer *cast_indexer(dft_indexer_handle_t indexer) {
 
 dft_indexer_handle_t dft_indexer_create(const char *gz_path,
                                         const char *idx_path,
-                                        size_t checkpoint_size,
+                                        uint64_t checkpoint_size,
                                         int force_rebuild) {
     if (!gz_path || !idx_path || checkpoint_size == 0) {
         DFTRACER_UTILS_LOG_ERROR("Invalid parameters for indexer creation", "");
